fix: Refresh dropdown when image is added
